Add files via upload
Description:
The dataset in file Restaurant tips dataset.xlsx contains tips data for different customers. The following are the features in the dataset:
sex	Gender of the customer
smoker	Indicates if the customer is a smoker or not
day	Day of the restaurant visit
time	Indicates whether the tip was for lunch or dinner
size	Number of members dining
total bill	Bill amount in USD
tip	Tip amount in USD
The following project tasks are required to be performed in excel:
•	Use the restaurant tips file for the analytics using Excel
•	Find out if there are any missing values and clean the data
•	Find the features that are independent and dependent
•	Identify which predictive problem is needed.
•	Encode the categorical variables to numeric values using IF conditions
•	Build an appropriate model with the dataset. 
•	Calculate the predicted and actual tips values.
•	Calculate the RMSE(Root Mean Square Error) of the model. RMSE is root of mean of square errors.
 
Tools required: Microsoft Excel, Data Analysis Add-in.
 
Expected Deliverables:  Model to predict restaurant tips given input values with the mathematical equation for predicting the tips value.